Revert "refactor(core): some runtime methods should live on the module map (#19502)"
This reverts commit 88e6e9c1e6263f326d30b7fcc33465e0428faf33.
